Get title from INSEE series idbank

Get title from INSEE series idbank

get_insee_title(..., lang = "en")

Arguments

...
list of series key (idbank)

lang
language of the title, by default it is Engligh, if lang is different from "en" then French will be the title's language

Value

a character vector with the titles

Details

Query INSEE website to get series title from series key (idbank). Any query to INSEE database can handle around 400 idbanks at maximum, if necessary the idbank list will then be splitted in several lists of 400 idbanks each. Consequently, it is not advised to use it on the whole idbank dataset, the user should filter the idbank dataset first.

Examples

#example 1 : industrial production index on manufacturing and industrial activities
title = get_insee_title("010537900")

#> 1/1 - Data download :

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> 1/1 - Dataframe build : 100%

#> 
#> Data cached : C:\Users\XLAPDO\AppData\Local\Temp\RtmpwzNSBn/8c931ff633c5c50d595e5c3d4503773a.rds
#> 


# \donttest{
#example 2 : automotive industry and overall industrial production
library(tidyverse)

idbank_list = get_idbank_list()

#> Cached data has been used


idbank_list_selected =
 idbank_list %>%
 filter(nomflow == "IPI-2015") %>% #industrial production index dataset
 filter(dim1 == "M") %>% #monthly
 filter(dim5 == "INDICE") %>% #index
 filter(dim8  == "CVS-CJO") %>% #Working day and seasonally adjusted SA-WDA
 filter(str_detect(dim4,"^29$|A10-BE")) %>% #automotive industry and overall industrial production
 mutate(title = get_insee_title(idbank))

#> 1/1 - Data download :

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> 1/1 - Dataframe build :

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> 
#> Data cached : C:\Users\XLAPDO\AppData\Local\Temp\RtmpwzNSBn/9c776c3056d832de34c758b60ed66cbe.rds
#> 

# }