Add metadata to the raw data

Add metadata to the raw data

add_insee_metadata(df)

Arguments

df
a dataframe containing data obtained from get_insee_idbank or get_insee_dataset

Value

a tibble with the data given as parameter plus the corresponding metadata

Details

Add metadata to the raw data obtained from get_insee_idbank or get_insee_dataset

Examples

# \donttest{
library(tidyverse)

#> ── Attaching packages ─────────────────────────────────────── tidyverse 1.3.0 ──

#> ✔ ggplot2 3.3.3     ✔ purrr   0.3.4
#> ✔ tibble  3.0.5     ✔ dplyr   1.0.3
#> ✔ tidyr   1.1.2     ✔ stringr 1.4.0
#> ✔ readr   1.4.0     ✔ forcats 0.5.0

#> ── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
#> ✖ dplyr::filter() masks stats::filter()
#> ✖ dplyr::lag()    masks stats::lag()

data =
 get_insee_idbank("001694061") %>%
 add_insee_metadata()

#> 1/1 - Data download & Dataframe build

#> Data cached : /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpIVBrnA/fcdad636cc51d27630f60fbfe871d094.rds
#> 

#> Downloading: 550 B     Downloading: 550 B     Downloading: 1.9 kB     Downloading: 1.9 kB     Downloading: 2.1 kB     Downloading: 2.1 kB     Downloading: 3 kB     Downloading: 3 kB     Downloading: 4.5 kB     Downloading: 4.5 kB     Downloading: 5.4 kB     Downloading: 5.4 kB     Downloading: 5.9 kB     Downloading: 5.9 kB     Downloading: 6 kB     Downloading: 6 kB     Downloading: 6 kB     Downloading: 6 kB     Downloading: 7.4 kB     Downloading: 7.4 kB     Downloading: 8.9 kB     Downloading: 8.9 kB     Downloading: 9.7 kB     Downloading: 9.7 kB     Downloading: 11 kB     Downloading: 11 kB     Downloading: 11 kB     Downloading: 11 kB     Downloading: 13 kB     Downloading: 13 kB     Downloading: 13 kB     Downloading: 13 kB     Downloading: 14 kB     Downloading: 14 kB     Downloading: 14 kB     Downloading: 14 kB       |                                                                              |                                                                      |   0%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |                                                                      |   1%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=                                                                     |   1%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=                                                                     |   2%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==                                                                    |   3%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===                                                                   |   4%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===                                                                   |   5%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |====                                                                  |   5%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |====                                                                  |   6%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=====                                                                 |   7%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=====                                                                 |   8%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |======                                                                |   8%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |======                                                                |   9%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=======                                                               |   9%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=======                                                               |  10%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |========                                                              |  11%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |========                                                              |  12%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=========                                                             |  12%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=========                                                             |  13%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==========                                                            |  14%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==========                                                            |  15%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===========                                                           |  15%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===========                                                           |  16%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |============                                                          |  17%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |============                                                          |  18%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=============                                                         |  18%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=============                                                         |  19%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==============                                                        |  20%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===============                                                       |  21%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===============                                                       |  22%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |================                                                      |  22%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |================                                                      |  23%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=================                                                     |  24%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=================                                                     |  25%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==================                                                    |  25%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==================                                                    |  26%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===================                                                   |  27%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |====================                                                  |  28%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |====================                                                  |  29%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=====================                                                 |  30%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=====================                                                 |  31%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |======================                                                |  31%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |======================                                                |  32%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=======================                                               |  33%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |========================                                              |  34%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |========================                                              |  35%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=========================                                             |  35%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=========================                                             |  36%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==========================                                            |  37%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==========================                                            |  38%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===========================                                           |  38%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===========================                                           |  39%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |============================                                          |  39%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |============================                                          |  40%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=============================                                         |  41%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=============================                                         |  42%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==============================                                        |  42%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==============================                                        |  43%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==============================                                        |  44%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===============================                                       |  44%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===============================                                       |  45%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |================================                                      |  45%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |================================                                      |  46%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=================================                                     |  47%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=================================                                     |  48%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==================================                                    |  48%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==================================                                    |  49%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===================================                                   |  50%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |====================================                                  |  51%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |====================================                                  |  52%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=====================================                                 |  52%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=====================================                                 |  53%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |======================================                                |  54%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |======================================                                |  55%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=======================================                               |  55%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=======================================                               |  56%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |========================================                              |  56%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |========================================                              |  57%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |========================================                              |  58%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=========================================                             |  58%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=========================================                             |  59%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==========================================                            |  60%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==========================================                            |  61%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===========================================                           |  61%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===========================================                           |  62%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |============================================                          |  62%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |============================================                          |  63%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=============================================                         |  64%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=============================================                         |  65%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==============================================                        |  65%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==============================================                        |  66%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===============================================                       |  67%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |================================================                      |  68%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |================================================                      |  69%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=================================================                     |  69%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=================================================                     |  70%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==================================================                    |  71%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==================================================                    |  72%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===================================================                   |  73%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |====================================================                  |  74%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |====================================================                  |  75%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=====================================================                 |  75%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=====================================================                 |  76%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |======================================================                |  77%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |======================================================                |  78%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=======================================================               |  78%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=======================================================               |  79%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |========================================================              |  80%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=========================================================             |  81%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=========================================================             |  82%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==========================================================            |  82%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==========================================================            |  83%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===========================================================           |  84%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===========================================================           |  85%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |============================================================          |  85%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |============================================================          |  86%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=============================================================         |  87%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=============================================================         |  88%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==============================================================        |  88%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==============================================================        |  89%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===============================================================       |  90%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===============================================================       |  91%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |================================================================      |  91%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |================================================================      |  92%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=================================================================     |  92%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |=================================================================     |  93%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==================================================================    |  94%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |==================================================================    |  95%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===================================================================   |  95%

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===================================================================   |  96%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |====================================================================  |  97%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===================================================================== |  98%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |===================================================================== |  99%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |======================================================================|  99%

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>   |                                                                              |======================================================================| 100%

#> Cached data has been used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

# }