Add a title column to the idbank list dataset

Add a title column to the idbank list dataset

add_insee_title(df, n_split, lang = "en", split = TRUE, clean = TRUE)

Arguments

df
a dataframe containing an idbank column called "idbank" or "IDBANK"

n_split
number of new columns, by default the maximum is chosen

lang
returns an English title, by default is "en", any other value returns a French title

split
split the title column in several columns, by default is TRUE

clean
remove the columns filled with NA (missing value), by default is TRUE

Value

the same dataframe but with one or several title columns

Details

this function uses extensively the get_insee_title function. Then, it should be used on an already filtered dataset, not on the full idbank dataset (cf. get_insee_title). The number of separators in the official INSEE title can vary and is not normalized. Beware all title columns created may not be a cleaned dimension label.

Examples

# \donttest{
library(tidyverse)

#> ── Attaching packages ─────────────────────────────────────── tidyverse 1.3.0 ──

#> ✔ ggplot2 3.3.2     ✔ purrr   0.3.4
#> ✔ tibble  3.0.3     ✔ dplyr   1.0.2
#> ✔ tidyr   1.1.2     ✔ stringr 1.4.0
#> ✔ readr   1.3.1     ✔ forcats 0.5.0

#> ── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
#> ✖ dplyr::filter() masks stats::filter()
#> ✖ dplyr::lag()    masks stats::lag()

idbank_list = get_idbank_list() #idbank list

#> 
#> Data cached : /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpBojNFw/32f277c98ef9977b40bfc84e14a3df54.rds
#> 

idbank_empl = idbank_list %>%
 filter(nomflow == "EMPLOI-SALARIE-TRIM-NATIONAL") %>% #employment
 slice(1:15) %>%
 add_insee_title()

#> 1/1 - Data download & Dataframe build

#> Data cached : /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpBojNFw/74916d4fc0cb008e77c1c831e699e673.rds
#> 

# }